#783 Performance issue when loading many files with SFTPMessageProvider
If the duplicate check is enabled then there is a performance issue with the SFTPConsumer (the SFTPMessageProvider and related classes).

Basically if the consumer wants to check for a new file and it first results in a list of x files, then it goes through all these files and calls an sql statement to detect if the file is a duplicate. Then it returns the whole filtered lists which gets sorted chronologically later (if enabled). Then it only takes the first file.

Now for the next file the whole process begins again. So e.g. if 4000 files should be consumed (eager = true), to find each next file to process all 4000 files are checked in the DB again and again.

This leads to a 11h of processing time e.g. in my tests with 4000 files, compared to about 1h after the fix.

The fix consists of
- not creating unnecessary String for disabled debug logging
- Allowing to cache the duplicate file info by making a Cache configurable to the HibernateBaseFileTransferDaoImpl
- Sort the files if needed as early as possible, so that we can skip unnecessary processing and directly return the first element, if only one is needed (which is currently the case anyway, as that get(0) is hard coded